Win32 updates, mainly using VS 2005 instead of VS 6.
git-svn-id: https://vdsf.svn.sourceforge.net/svnroot/vdsf/trunk@881 53c26a9f-4021-0410-9f73-c4284c400f72

+++ src/Common/localtime_r.c
@@ -245,8 +245,8 @@ timesub( const time_t * const		timep,
 			break;
 		}
 	}
-	days = *timep / SECSPERDAY;
-	rem = *timep % SECSPERDAY;
+	days = (long)*timep / SECSPERDAY;
+	rem  = (long)*timep % SECSPERDAY;
 #ifdef mc68k
 	if (*timep == 0x80000000) {
 		/*
